+++ pomdp_py/__init__.py
@@ -1,5 +1,4 @@
 from pomdp_py.utils import util
-from pomdp_py.utils.interfaces import conversion
 from pomdp_py.framework.basics import *
 from pomdp_py.framework.oopomdp import *
 from pomdp_py.framework.planner import *
@@ -17,3 +16,6 @@ from pomdp_py.algorithms.bsp.blqr import BLQR
 from pomdp_py.algorithms.visual import visual
 from pomdp_py.representations.belief.histogram import update_histogram_belief
 from pomdp_py.representations.belief.particles import update_particles_belief
+
+from pomdp_py.utils.interfaces.conversion import to_pomdp_file, to_pomdpx_file
+from pomdp_py.utils.interfaces.solvers import vi_pruning