+++ pomdp_py/utils/interfaces/tests/test_local.sh
@@ -1,3 +1,5 @@
 python test_conversion_pomdp-solve.py ~/software/pomdp-solve-5.4/src/pomdp-solve
 
 python test_conversion_pomdpx.py ~/software/sarsop/src/pomdpconvert
+
+python test_vi_pruning.py ~/software/pomdp-solve-5.4/src/pomdp-solve